Postgres v0.7 reach dead end code path - add database tables, repos, and services - add DataWriter class - LIDL full db ingestion path working
Notes: As with the previous prostgres experiment, data ingestion via
Querys is proving impractical(takes a long time and 100% CPU thread
while doing it). This time I used a different database
design and Spring Data JPA intead of JDBC with undervelming results.
This aproach is now DEAD, I will keep the database but attempt to do
file based ingestion which has already proven to work, but I can't use
the last implementation since the database design is to different.

+++ src/main/java/com/brunovelcl/pricetracker/DataParser/entities/ParsedValuesBuilder.java
@@ -2,10 +2,12 @@ package com.brunovelcl.pricetracker.DataParser.entities;
 
 import com.brunovelcl.pricetracker.database.entities.Stores;
 
+import java.math.BigDecimal;
+
 public class ParsedValuesBuilder {
 
     private Long barcode;
-    private Double price;
+    private BigDecimal price;
     private String productName;
     private String brand;
     private String unit_quantity;
@@ -19,7 +21,7 @@ public class ParsedValuesBuilder {
         this.barcode = barcode;
     }
 
-    public void price(Double price) {
+    public void price(BigDecimal price) {
         this.price = price;
     }
